43. Test python socketio & uvicorn/source_code/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
bidict==0.21.4
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
frozenlist==1.2.0
h11==0.12.0
idna==3.3
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pydantic==1.8.2
python-engineio==4.3.0
python-socketio==5.5.0
requests==2.26.0
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
websocket-client==1.2.3
yarl==1.7.2